[feature/admin-controller] fixed genre

+++ src/Controller/GenreController.php
@@ -2,19 +2,15 @@
 
 namespace Controller;
 
-use Exception;
 use Framework\Validator\GenreValidator;
 use Entity\GenreEntity;
 
 class GenreController extends AbstractController {
-
     /**
      * Shows genre list
      * @return array
      */
-    public function ShowGenreList()
-    {
-
+    public function ShowGenreList() {
         $genreRepository = $this->getRepository('genre');
         $genreList = $genreRepository->loadAll();
         $context = [
@@ -22,35 +18,25 @@ class GenreController extends AbstractController {
         ];
         return $this->render('genre', $context);
     }
-
-
     /**
      * add a new genre name in genre list
      */
     public function addGenre() {
-
-        $validator = new \Entity\GenreValidator;
-
+        $validator = new GenreValidator;
         // build properties array 
         $properties = [
-            'name' => $this->getPostParam('name')
+            'name' => $this->getPostParam('genreName')
         ];
-
         // build an entity 
         $genre = new \Entity\GenreEntity($properties);
-
         $genreName = $genre->getName();
         $errors = $validator->validate($genre);
-
-
         if (!empty($errors)) {
             $this->addErrorMessage($errors);
             return $this->render('genre', ['last_name' => $this->request->get('name')]);
         }
-
         // get the repository
         $genreRepository = $this->getRepository('genre');
-
         // check if genre name exists in db
         try {
             $genreByName = $genreRepository->loadByProperties(['name' => $genreName]);
@@ -59,31 +45,25 @@ class GenreController extends AbstractController {
             $this->addErrorMessage('We\'re sorry, something went terribly wrong while trying to add the genre name. Please try again later.'); // ? 
             return $this->render('genre');
         }
-
         if (count($genreByName) !== 0) {
             $this->addErrorMessage('This genre name is already associated with another name.');
             return $this->render('genre', ['last_name' => $this->request->get('name')]);
         }
-
         // add to db
-
         try {
             $genreRepository->save($genre);
         } catch (\Exception $ex) {
             $this->addErrorMessage('We\'re sorry, something went terribly wrong while trying to add the genre name. Please try again later.'); // ??
             return $this->render('genre');
         }
-
         $this->addSuccessMessage('Genre name succesfully addeed!');
         return $this->redirectRoute('admin_genre_show_all');
     }
-
     /*
      * delete a genre name from genre list
      */
-
-    public function deleteGenre()
-    {
+    public function deleteGenre() {
+        
          // get the repository
         $genreRepository = $this->getRepository('genre');
      
@@ -92,27 +72,32 @@ class GenreController extends AbstractController {
             'name' => $this->getPostParam('name')
           
         ];
-
+        
         $idGenre=$this->getCustomParam('id');
-
         $genres=$genreRepository->loadByProperties(['id'=>$idGenre]);
         
-        //testezi de null
-        $genre = reset($genres);
+        //check if the id is empty
+        if (empty($genres)){
+             
+            $this->addErrorMessage('The id doesn\'t exist');
+            return $this->render('genre');
         
+        }
+        
+        $genre = reset($genres);
+       
          try {
             $genreRepository->delete($genre);
         } catch (\Exception $ex) {
             $this->addErrorMessage('We\'re sorry, something went terribly wrong while trying to delete the genre name. Please try again later.'); // ??
             return $this->render('genre');
         }
-
         $this->addSuccessMessage('Genre name succesfully deleted!');
         return $this->redirectRoute('admin_genre_show_all');
         
     }
-
-    /*
+    
+     /*
      * edit a genre name from genr elist
      */
 
@@ -151,13 +136,10 @@ class GenreController extends AbstractController {
         $successResponse['message'] = 'Updated!';
         $successResponse['title'] = 'Success!';
         $successResponse['type'] = 'success';
-        
         return $this->application->json($successResponse);
     }
 
-    public function getClassName()
-    {
+    public function getClassName() {
         return 'Controller\\GenreController';
     }
-
 }
